[Bugfix]: added include\exclude [Feature]: refactored structure
