--- src/feature/echo/logic/controller.ts
@@ -1,5 +1,0 @@
-import { Bot } from '../../../util/custom-types/telegraf.types';
-
-export interface Controller {
-  register(bot: Bot): void;
-}
