+++ src/logic/controller.ts
@@ -1,0 +1,5 @@
+import { Bot } from '../util/custom-types/telegraf.types';
+
+export interface Controller {
+  register(bot: Bot): void;
+}
